% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appl2.R
\name{appl2.label.as.vector}
\alias{appl2.label.as.vector}
\title{Converts a label of the format X...-Y where X is the substance abbreviation and Y is the concentration. CTRL is treated as abbreviation for control.}
\usage{
appl2.label.as.vector(label)
}
\description{
Converts a label of the format X...-Y where X is the substance abbreviation and Y is the concentration. CTRL is treated as abbreviation for control.
}
