% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appl2.R
\name{appl2.plot.dose.response}
\alias{appl2.plot.dose.response}
\title{TBD}
\usage{
appl2.plot.dose.response(
  aggregated.data,
  model.data,
  predicted.data,
  substance.name,
  x.upper.lim = 100
)
}
\description{
TBD
}
