% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.processing.R
\name{cap.normalized.data}
\alias{cap.normalized.data}
\title{Cleans a dataframe of normalized data from values bigger than 1.}
\usage{
cap.normalized.data(data, mode, normal.colname = "response")
}
\arguments{
\item{mode}{What to do with data bigger 1 (1 - is capped at 1, 2- is set to NA, default - do nothing)}
}
\description{
Cleans a dataframe of normalized data from values bigger than 1.
}
\note{
It is persumed, that control data is marked by teh entry 'control' in a column named 'substance'
}
