% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dose.response.plot}
\alias{dose.response.plot}
\title{TBD}
\usage{
dose.response.plot(
  aggregated.data,
  model.data,
  title,
  x.upper.lim,
  xlab,
  ylab,
  filename,
  no.conf = FALSE
)
}
\description{
TBD
}
