% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.processing.R
\name{filter.data.by.substances}
\alias{filter.data.by.substances}
\title{TBD}
\usage{
filter.data.by.substances(
  data,
  substance.names,
  substance.colname = "substance",
  concentration.colname = "concentration",
  include.control = TRUE,
  concentrations.factors = TRUE
)
}
\description{
TBD
}
