% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.analysis.R
\name{fit.exp.model}
\alias{fit.exp.model}
\title{TBD}
\usage{
fit.exp.model(
  data,
  response.variable.name = "response",
  guesses = c(a = 1),
  lower = c(a = 0),
  print.summary = TRUE,
  algorithm = "default",
  trace = FALSE,
  control = nls.control()
)
}
\description{
TBD
}
