% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.analysis.R
\name{fit.generic.model}
\alias{fit.generic.model}
\title{TBD}
\usage{
fit.generic.model(
  equation,
  data,
  response.variable.name,
  guesses,
  lower,
  print.summary,
  algorithm,
  trace,
  control
)
}
\description{
TBD
}
