% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.analysis.R
\name{fit.sq.model}
\alias{fit.sq.model}
\title{TBD}
\usage{
fit.sq.model(
  data,
  response.variable.name = "response",
  guesses = list(a = 1, b = 1, c = 1),
  lower = list(a = NULL, b = 0, c = 0),
  print.summary = TRUE,
  algorithm = "default",
  trace = FALSE,
  control = nls.control()
)
}
\description{
TBD
}
