% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.analysis.R
\name{normalize.data}
\alias{normalize.data}
\title{TBD. If use.normalize.value is FALSE, certain rows out of data are selected and the mean of the values in their colname.normalize column is used as normalize.value.}
\usage{
normalize.data(
  data,
  colname.normalize,
  colname.selection,
  object.selection,
  normalize.value = 1,
  use.normalize.value = FALSE
)
}
\arguments{
\item{colname.normalize}{The name of the column in which the data, data should be normalized are located. Further if use.normalize.value is false, this column will be used to calculate normalize.value.}

\item{colname.selection}{The name of the column in which the object.selection object should be matched.}

\item{object.selection}{The object which must be matched in the column specified by colname.selection for rows to be included in the normalize.value.}

\item{normalize.value}{The value which to use for normalizing. Is ignored if use.normalize.value is FALSE.}

\item{use.normalize.value}{A boolean to indicate whether or not the normalize.value parameter should be used.}
}
\description{
TBD. If use.normalize.value is FALSE, certain rows out of data are selected and the mean of the values in their colname.normalize column is used as normalize.value.
}
